+++ pull_request_template.md
@@ -4,8 +4,8 @@
 
 ### Чек-лист
 
-- [ ] ПР имеет четкое **название**, короткое и описывающее его суть.
-- [ ] ПР имеет четкое **описание** того, что в рамках него было сделано.
-- [ ] Я провел **само-ревью** ПРа. Там нет кода, который я забыл удалить, случайно закоммитил и т.п.
-- [ ] ПР не содержит кода, который не соответствует цели ПРа.
-- [ ] ПР не содержит мерж конфликтов.
+- [] ПР имеет четкое **название**, короткое и описывающее его суть.
+- [] ПР имеет четкое **описание** того, что в рамках него было сделано.
+- [] Я провел **само-ревью** ПРа. Там нет кода, который я забыл удалить, случайно закоммитил и т.п.
+- [] ПР не содержит кода, который не соответствует цели ПРа.
+- [] ПР не содержит мерж конфликтов.
