--- src/app/providers/app-router/ui/AppRouter.module.scss
@@ -1,3 +1,0 @@
-.pageWrapper {
-  height: 100%;
-}
