+++ src/app/providers/app-router/ui/app-router.module.scss
@@ -1,0 +1,4 @@
+.pageWrapper {
+    height: 100%;
+  }
+
