+++ src/app/types/modules.d.ts
@@ -1,0 +1,11 @@
+declare module 'react-manifest';
+
+declare module '*.scss' {
+  type IClassNames = Record<string, string>;
+  const classNames: IClassNames;
+  export = classNames;
+}
+
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.jpeg';
