--- src/repositories/post-card-list/api/types.ts
@@ -1,5 +1,0 @@
-import { type PostCard } from '../../../models/post-card-list';
-
-export interface GetPostCardApiDto {
-    postCards: PostCard[]
-}
