--- src/repositories/post-card-list/store/index.ts
@@ -1,13 +1,0 @@
-import { type PostCard } from '../../../models/post-card-list';
-
-export class PostStoreRepository {
-    // TODO взаимодействие с Redux здесь (предположительно)
-
-    public readModel (): PostCard[] {
-        return [];
-    }
-
-    public saveModel (postCardList: PostCard[]): void {
-
-    }
-}
